allow overriding stop_at_empty when setting range, move parsers into separate files
