+++ CMakeLists.txt
@@ -14,7 +14,11 @@ include_directories(src/excel/numformat/include)
 include_directories(src/excel/include)
 add_subdirectory(src/excel/numformat)
 
-set(EXTENSION_SOURCES src/excel/excel_extension.cpp src/excel/xlsx/zip_file.cpp src/excel/xlsx/read_xlsx.cpp src/excel/xlsx/write_xlsx.cpp)
+set(EXTENSION_SOURCES
+        src/excel/excel_extension.cpp
+        src/excel/xlsx/zip_file.cpp
+        src/excel/xlsx/read_xlsx.cpp
+        src/excel/xlsx/write_xlsx.cpp)
 
 build_static_extension(${TARGET_NAME} ${EXTENSION_SOURCES}
                        ${NUMFORMAT_OBJECT_FILES})
