+++ src/excel/xlsx/read_xlsx.cpp
@@ -3,9 +3,15 @@
 #include "xlsx/xml_parser.hpp"
 #include "xlsx/zip_file.hpp"
 #include "xlsx/xlsx_parts.hpp"
+#include "xlsx/string_table.hpp"
 
-#include "duckdb/common/string_map_set.hpp"
-#include "duckdb/common/types/string_type.hpp"
+#include "xlsx/parsers/relationship_parser.hpp"
+#include "xlsx/parsers/content_types_parser.hpp"
+#include "xlsx/parsers/stylesheet_parser.hpp"
+#include "xlsx/parsers/shared_strings_parser.hpp"
+#include "xlsx/parsers/workbook_parser.hpp"
+#include "xlsx/parsers/worksheet_parser.hpp"
+
 #include "duckdb/common/types/time.hpp"
 #include "duckdb/function/table_function.hpp"
 #include "duckdb/main/database.hpp"
@@ -13,1124 +19,9 @@
 #include "duckdb/parser/expression/constant_expression.hpp"
 #include "duckdb/parser/expression/function_expression.hpp"
 #include "duckdb/parser/tableref/table_function_ref.hpp"
-#include "duckdb/storage/arena_allocator.hpp"
 #include "duckdb/main/query_result.hpp"
 
 namespace duckdb {
-
-static const auto WBOOK_CONTENT_TYPE = "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet.main+xml";
-static const auto SHEET_CONTENT_TYPE = "application/vnd.openxmlformats-officedocument.spreadsheetml.worksheet+xml";
-
-//-------------------------------------------------------------------
-// "[Content_Types].xml" Parser
-//-------------------------------------------------------------------
-struct ContentInfo {
-	string wbook_path;
-	string sheet_path;
-};
-
-class ContentParser final : public XMLParser {
-public:
-	static ContentInfo ParseContentTypes(ZipFileReader &stream) {
-		ContentParser parser;
-		parser.ParseAll(stream);
-		return std::move(parser.info);
-	}
-protected:
-	void OnStartElement(const char *name, const char **atts) override {
-		switch(state) {
-		case State::START:
-			if(MatchTag("Types", name)) {
-				state = State::TYPES;
-			} break;
-		case State::TYPES:
-			if(MatchTag("Override", name)) {
-				state = State::OVERRIDE;
-
-				// Now, also extract the attributes
-				const char *ctype = nullptr;
-				const char *pname = nullptr;
-
-				for(idx_t i = 0; atts[i]; i += 2) {
-					if(strcmp(atts[i], "ContentType") == 0) {
-						ctype = atts[i + 1];
-					} else if(strcmp(atts[i], "PartName") == 0) {
-						pname = atts[i + 1];
-					}
-				}
-
-				if(ctype && pname) {
-					if(strcmp(ctype, WBOOK_CONTENT_TYPE) == 0) {
-						info.wbook_path = pname;
-					}
-					else if(strcmp(ctype, SHEET_CONTENT_TYPE) == 0) {
-						info.sheet_path = pname;
-					}
-				} else {
-					throw InvalidInputException("Invalid content type entry in [Content_Types].xml");
-				}
-
-			} break;
-		default:
-			break;
-		}
-	}
-	void OnEndElement(const char *name) override {
-		switch(state) {
-		case State::OVERRIDE:
-			if(MatchTag("Override", name)) {
-				state = State::TYPES;
-			} break;
-		case State::TYPES:
-			if(MatchTag("Types", name)) {
-				state = State::END;
-				Stop(false);
-			} break;
-		default: break;
-		}
-	}
-private:
-	enum class State : uint8_t { START, TYPES, OVERRIDE, END };
-	ContentInfo info;
-	State state = State::START;
-};
-
-//-------------------------------------------------------------------
-// "xl/workbook.xml" Parser
-//-------------------------------------------------------------------
-
-class WorkBookParser final : public XMLParser {
-public:
-	static vector<pair<string, string>> GetSheets(ZipFileReader &stream) {
-		WorkBookParser parser;
-		parser.ParseAll(stream);
-		return std::move(parser.sheets);
-	}
-private:
-	void OnStartElement(const char *name, const char **atts) override {
-		switch(state) {
-		case State::START:
-			if(MatchTag("workbook", name)) {
-				state = State::WORKBOOK;
-			} break;
-		case State::WORKBOOK:
-			if(MatchTag("sheets", name)) {
-				state = State::SHEETS;
-			} break;
-		case State::SHEETS:
-			if(MatchTag("sheet", name)) {
-				state = State::SHEET;
-				// Now extract attributes
-				const char *sheet_name = nullptr;
-				const char *sheet_ridx = nullptr;
-				for(idx_t i = 0; atts[i]; i += 2) {
-					if(strcmp(atts[i], "name") == 0) {
-						sheet_name = atts[i + 1];
-					} else if(strcmp(atts[i], "r:id") == 0) {
-						sheet_ridx = atts[i + 1];
-					}
-				}
-				if(sheet_name && sheet_ridx) {
-					sheets.emplace_back(sheet_name, sheet_ridx);
-				} else {
-					throw InvalidInputException("Invalid sheet entry in workbook.xml");
-				}
-			} break;
-		default:
-			break;
- 		}
-	}
-	void OnEndElement(const char *name) override {
-		switch(state) {
-		case State::SHEET:
-			if(MatchTag("sheet", name)) {
-				state = State::SHEETS;
-			} break;
-		case State::SHEETS:
-			if(MatchTag("sheets", name)) {
-				state = State::WORKBOOK;
-			} break;
-		case State::WORKBOOK:
-			if(MatchTag("workbook", name)) {
-				Stop(false);
-			} break;
-		default:
-			break;
-		}
-	}
-private:
-	enum class State { START, WORKBOOK, SHEETS, SHEET, };
-	State state = State::START;
-	vector<pair<string, string>> sheets;
-};
-
-//-------------------------------------------------------------------
-// Relationship Parser
-//-------------------------------------------------------------------
-
-struct XLSXRelation {
-	string id;
-	string type;
-	string target;
-};
-
-class RelParser final : public XMLParser {
-public:
-	static vector<XLSXRelation> ParseRelations(ZipFileReader &stream) {
-		RelParser parser;
-		parser.ParseAll(stream);
-		return std::move(parser.relations);
-	}
-
-protected:
-	void OnStartElement(const char *name, const char **atts) override {
-		switch(state) {
-		case State::START:
-			if(MatchTag("Relationships", name)) {
-				state = State::RELATIONSHIPS;
-			}
-			break;
-		case State::RELATIONSHIPS:
-			if(MatchTag("Relationship", name)) {
-				state = State::RELATIONSHIP;
-
-				// Extract the attributes
-				const char *rid = nullptr;
-				const char *rtype = nullptr;
-				const char *rtarget = nullptr;
-
-				for(idx_t i = 0; atts[i]; i += 2) {
-					if(strcmp(atts[i], "Id") == 0) {
-						rid = atts[i + 1];
-					} else if(strcmp(atts[i], "Type") == 0) {
-						rtype = atts[i + 1];
-					} else if(strcmp(atts[i], "Target") == 0) {
-						rtarget = atts[i + 1];
-					}
-				}
-
-				if(rid && rtype && rtarget) {
-					relations.emplace_back(XLSXRelation{rid, rtype, rtarget});
-				} else {
-					throw InvalidInputException("Invalid relationship entry in _rels/.rels");
-				}
-			}
-			break;
-		default:
-			break;
-		}
-	}
-	void OnEndElement(const char *name) override {
-		switch(state) {
-		case State::RELATIONSHIP:
-			if(MatchTag("Relationship", name)) {
-				state = State::RELATIONSHIPS;
-			}
-			break;
-		case State::RELATIONSHIPS:
-			if(MatchTag("Relationships", name)) {
-				Stop(false);
-			}
-			break;
-		default: break;
-		}
-	}
-private:
-	enum class State : uint8_t { START, RELATIONSHIPS, RELATIONSHIP };
-	State state = State::START;
-
-	vector<XLSXRelation> relations;
-};
-
-//-------------------------------------------------------------------
-// StyleSheet Parser
-//-------------------------------------------------------------------
-
-class XLSXStyleParser final : public XMLParser {
-public:
-	unordered_map<idx_t, LogicalType> number_formats;
-	vector<LogicalType> cell_styles;
-
-protected:
-	void OnStartElement(const char *name, const char **atts) override;
-	void OnEndElement(const char *name) override;
-private:
-	enum class State : uint8_t { START, STYLESHEET, NUMFMTS, NUMFMT, CELLXFS, XF };
-	State state = State::START;
-};
-
-template <class... ARGS>
-static bool StringContainsAny(const char *str, ARGS &&...args) {
-	for (auto &&substr : {args...}) {
-		if (strstr(str, substr) != nullptr) {
-			return true;
-		}
-	}
-	return false;
-}
-
-void XLSXStyleParser::OnStartElement(const char *name, const char **atts) {
-	switch(state) {
-	case State::START:
-		if(MatchTag("styleSheet", name)) {
-			state = State::STYLESHEET;
-		}
-		break;
-	case State::STYLESHEET:
-		if(MatchTag("numFmts", name)) {
-			state = State::NUMFMTS;
-		} else if(MatchTag("cellXfs", name)) {
-			state = State::CELLXFS;
-		}
-		break;
-	case State::NUMFMTS: {
-		state = State::NUMFMT;
-
-		const char* id_ptr = nullptr;
-		const char* format_ptr = nullptr;
-
-		for(idx_t i = 0; atts[i]; i += 2) {
-			if(strcmp(atts[i], "numFmtId") == 0) {
-				id_ptr = atts[i + 1];
-			} else if(strcmp(atts[i], "formatCode") == 0) {
-				format_ptr = atts[i + 1];
-			}
-		}
-		if(!id_ptr) {
-			throw InvalidInputException("Invalid numFmt entry in styles.xml");
-		}
-		const auto id = strtol(id_ptr, nullptr, 10);
-		if(id <= 163 || format_ptr == nullptr) {
-			break;
-		}
-
-		const auto has_date_part = StringContainsAny(format_ptr, "DD", "dd", "YY", "yy");
-		const auto has_time_part = StringContainsAny(format_ptr, "HH", "hh", "h", "H");
-
-		if (has_date_part && has_time_part) {
-			number_formats.emplace(id, LogicalType::TIMESTAMP);
-		} else if (has_date_part) {
-			number_formats.emplace(id, LogicalType::DATE);
-		} else if (has_time_part) {
-			number_formats.emplace(id, LogicalType::TIME);
-		} else {
-			// If we dont know how to handle the format, default to the numeric value.
-			number_formats.emplace(id, LogicalType::DOUBLE); // TODO: Or double?
-		}
-
-	} break;
-	case State::CELLXFS: {
-		state = State::XF;
-		const char* id_ptr = nullptr;
-
-		for(idx_t i = 0; atts[i]; i += 2) {
-			if(strcmp(atts[i], "numFmtId") == 0) {
-				id_ptr = atts[i + 1];
-			}
-		}
-		if(!id_ptr) {
-			throw InvalidInputException("Invalid xf entry in styles.xml");
-		}
-		const auto id = strtol(id_ptr, nullptr, 10);
-		if (id < 164) {
-			// Special cases
-			if (id >= 14 && id <= 17) {
-				cell_styles.push_back(LogicalType::DATE);
-			} else if (id >= 18 && id <= 21) {
-				cell_styles.push_back(LogicalType::TIME);
-			} else if (id == 22) {
-				cell_styles.push_back(LogicalType::TIMESTAMP);
-			}
-		} else {
-			// Look up the ID in the format map
-			const auto it = number_formats.find(id);
-			if (it != number_formats.end()) {
-				cell_styles.push_back(it->second);
-			}
-		}
-	} break;
-	default:
-		break;
-	}
-}
-
-void XLSXStyleParser::OnEndElement(const char *name) {
-	switch(state) {
-		case State::NUMFMT:
-			if(MatchTag("numFmt", name)) {
-				state = State::NUMFMTS;
-			}
-			break;
-		case State::XF:
-			if(MatchTag("xf", name)) {
-				state = State::CELLXFS;
-			}
-			break;
-		case State::NUMFMTS:
-			if(MatchTag("numFmts", name)) {
-				state = State::STYLESHEET;
-			}
-			break;
-		case State::CELLXFS:
-			if(MatchTag("cellXfs", name)) {
-				state = State::STYLESHEET;
-			}
-			break;
-		case State::STYLESHEET:
-			if(MatchTag("styleSheet", name)) {
-				Stop(false);
-			}
-			break;
-		default:
-			break;
-	}
-}
-
-//-------------------------------------------------------------------
-// String Table
-//-------------------------------------------------------------------
-// StringTable is a simple class that stores a unique set of strings
-// in arena backed memory and allows fast access by index.
-
-class StringTable {
-public:
-	explicit StringTable(Allocator &alloc) : arena(alloc) {}
-
-	idx_t Add(const string_t &str);
-	const string_t& Get(idx_t val) const;
-	void Reserve(idx_t count);
-
-private:
-	ArenaAllocator			arena;
-	string_map_t<idx_t>		table;
-	vector<string_t>		index; // todo: make this a reference to the key in the map?
-};
-
-idx_t StringTable::Add(const string_t &str) {
-
-	// Check if the string is already in the map
-	const auto found = table.find(str);
-	if(found != table.end()) {
-		return found->second;
-	}
-
-	// Create a new entry
-	const auto val = index.size();
-
-	// If the string is inlined, just store it in the map and vector
-	if(str.IsInlined()) {
-		table[str] = val;
-		index.push_back(str);
-		return val;
-	}
-
-	// Otherwise copy the string to the arena
-	const auto len = str.GetSize();
-	const auto src = str.GetData();
-	const auto dst = arena.Allocate(len);
-	memcpy(dst, src, len);
-
-	const auto key = string_t(const_char_ptr_cast(dst), len);
-	table[key] = val;
-	index.push_back(key);
-
-	return val;
-}
-
-const string_t& StringTable::Get(const idx_t val) const {
-	return index[val];
-}
-
-void StringTable::Reserve(const idx_t count) {
-	table.reserve(count);
-	index.reserve(count);
-}
-
-//-------------------------------------------------------------------
-// String Table Parser
-//-------------------------------------------------------------------
-class SharedStringsParser : public XMLParser {
-protected:
-	virtual void OnUniqueCount(idx_t count) { }
-	virtual void OnString(const vector<char> &str) = 0;
-private:
-	void OnStartElement(const char *name, const char **atts) override {
-		switch(state) {
-		case State::START:
-			if(MatchTag("sst", name)) {
-				state = State::SST;
-				// Optionally look for the uniqueCount attributes
-				// TODO: Do we also look for count?
-				for(idx_t i = 0; atts[i]; i += 2) {
-					if(strcmp(atts[i], "uniqueCount") == 0) {
-						// TODO: Check that this succeeds!
-						const auto unique_count = atoi(atts[i + 1]);
-						OnUniqueCount(unique_count);
-					}
-				}
-			}
-			break;
-		case State::SST:
-			if(MatchTag("si", name)) {
-				state = State::SI;
-			}
-			break;
-		case State::SI:
-			if(MatchTag("t", name)) {
-				state = State::T;
-				// Enable text handling
-				EnableTextHandler(true);
-			}
-			break;
-		case State::T:
-			// TODO: Throw? We should never get here
-			break;
-		default:
-			// TODO: Count garbage iterations where we dont find anything and throw if we get too far
-			// to guard against corrupted files.
-			// we're probably going to have to set that in the parser base class itself
-		break;
-		}
-	}
-	void OnEndElement(const char *name) override {
-		switch(state) {
-		case State::T:
-			if(MatchTag("t", name)) {
-				// Disable text handling
-				EnableTextHandler(false);
-				state = State::SI;
-			}
-			break;
-		case State::SI:
-			if(MatchTag("si", name)) {
-				state = State::SST;
-				// Pass the string we've collected from the <t> tags to the handler
-				OnString(data);
-				// Reset the buffer
-				data.clear();
-			}
-			break;
-		case State::SST:
-			if(MatchTag("sst", name)) {
-				Stop(false);
-			}
-			break;
-		default:
-			break;
-		}
-	}
-	void OnText(const char *text, const idx_t len) override {
-		data.insert(data.end(), text, text + len);
-	}
-private:
-	enum class State : uint8_t { START, SST, SI, T };
-	State state = State::START;
-	vector<char> data;
-};
-
-// Parses the string table for a specific set of strings
-class StringTableSearcher final : public SharedStringsParser {
-public:
-	explicit StringTableSearcher(const vector<idx_t> &ids_p) : ids(ids_p) {
-		std::sort(ids.begin(), ids.end());
-	}
-
-	const unordered_map<idx_t, string> &GetResult() const {
-		return result;
-	}
-
-protected:
-	void OnString(const vector<char> &str) override {
-		if(current_idx >= ids.size()) {
-			// We're done, no more strings to find
-			Stop(false);
-			return;
-		}
-		const auto &id = ids[current_idx];
-		if(id == current_str) {
-			result[id] = string(str.data(), str.size());
-			current_idx++;
-		}
-		current_str++;
-	}
-private:
-	idx_t current_idx = 0;
-	idx_t current_str = 0;
-
-	vector<idx_t> ids;
-	unordered_map<idx_t, string> result;
-};
-
-// Parses the string table and populate it completely
-class StringTableParser final : public SharedStringsParser {
-public:
-	static void ParseStringTable(ZipFileReader &stream, StringTable &table) {
-		StringTableParser parser(table);
-		parser.ParseAll(stream);
-	}
-private:
-	explicit StringTableParser(StringTable &table_p) : table(table_p) {
-	}
-protected:
-	void OnString(const vector<char> &str) override {
-		table.Add(string_t(str.data(), str.size()));
-	}
-	void OnUniqueCount(const idx_t count) override {
-		table.Reserve(count);
-	}
-private:
-	StringTable &table;
-};
-
-//-------------------------------------------------------------------
-// SheetParser
-//-------------------------------------------------------------------
-
-class SheetParserBase : public XMLParser {
-public:
-	void OnText(const char *text, idx_t len) override;
-	void OnStartElement(const char *name, const char **atts) override;
-	void OnEndElement(const char *name) override;
-protected:
-	virtual void OnBeginRow(idx_t row_idx) {};
-	virtual void OnEndRow(idx_t row_idx) {};
-	virtual void OnCell(const XLSXCellPos &pos, XLSXCellType type, vector<char> &data, idx_t style) {}
-private:
-	enum class State : uint8_t { START, SHEETDATA, ROW, EMPTY_ROW, CELL, V, IS, T };
-	State state = State::START;
-
-	XLSXCellPos  cell_pos = {0, 0};
-	XLSXCellType cell_type = XLSXCellType::NUMBER;
-	vector<char> cell_data = {};
-	idx_t		 cell_style = 0;
-};
-
-void SheetParserBase::OnText(const char *text, idx_t len) {
-	if(cell_data.size() + len > XLSX_MAX_CELL_SIZE * 2) {
-		// Something is obviously wrong, error out!
-		throw InvalidInputException("XLSX: Cell data too large (is the file corrupted?)");
-	}
-	cell_data.insert(cell_data.end(), text, text + len);
-}
-
-void SheetParserBase::OnStartElement(const char *name, const char **atts) {
-	if(state == State::START && MatchTag("sheetData", name)) {
-		state = State::SHEETDATA;
-	}
-	else if(state == State::SHEETDATA && MatchTag("row", name)) {
-		state = State::ROW;
-
-		// Reset the column position
-		cell_pos.col = 0;
-
-		const char *rref_ptr = nullptr;
-		for(idx_t i = 0; atts[i]; i += 2) {
-			if(strcmp(atts[i], "r") == 0) {
-				rref_ptr = atts[i + 1];
-			}
-		}
-		// Default: Increment the row
-		if(!rref_ptr) {
-			cell_pos.row++;
-		} else {
-			// Else, jump to the row reference
-			// TODO: Verify this
-			cell_pos.row = strtol(rref_ptr, nullptr, 10);
-		}
-
-		OnBeginRow(cell_pos.row);
-	}
-	else if(state == State::ROW && MatchTag("c", name)) {
-		state = State::CELL;
-
-		// Reset the cell data
-		cell_data.clear();
-
-		// We're entering a cell. Parse the attributes
-		const char* type_ptr = nullptr;
-		const char* cref_ptr = nullptr;
-		const char* style_ptr = nullptr;
-		for(idx_t i = 0; atts[i]; i += 2) {
-			if(strcmp(atts[i], "t") == 0) {
-				type_ptr = atts[i + 1];
-			} else if (strcmp(atts[i], "r") == 0) {
-				cref_ptr = atts[i + 1];
-			} else if (strcmp(atts[i], "s") == 0) {
-				style_ptr = atts[i + 1];
-			}
-		}
-
-		// Default: 0
-		cell_style = style_ptr ? strtol(style_ptr, nullptr, 10) : 0;
-		// Default: NUMBER
-		cell_type = type_ptr ? ParseCellType(type_ptr) : XLSXCellType::NUMBER;
-		// Default: next cell
-		if(!cref_ptr) {
-			cell_pos.col++;
-		} else {
-			XLSXCellPos cref;
-			if(!cref.TryParse(cref_ptr)) {
-				throw InvalidInputException("Invalid cell reference in sheet");
-			}
-			if(cref.row != cell_pos.row) {
-				throw InvalidInputException("Cell reference does not match row reference in sheet");
-			}
-			cell_pos.col = cref.col;
-		}
-	}
-	else if(state == State::CELL && MatchTag("v", name)) {
-		state = State::V;
-		EnableTextHandler(true);
-	}
-	else if(state == State::CELL && MatchTag("is", name)) {
-		state = State::IS;
-	}
-	else if(state == State::IS && MatchTag("t", name)) {
-		state = State::T;
-		EnableTextHandler(true);
-	}
-}
-
-void SheetParserBase::OnEndElement(const char *name) {
-	if(state == State::SHEETDATA && MatchTag("sheetData", name)) {
-		Stop(false);
-	}
-	else if(state == State::ROW && MatchTag("row", name)) {
-		OnEndRow(cell_pos.row);
-		state = State::SHEETDATA;
-	}
-	else if(state == State::CELL && MatchTag("c", name)) {
-		OnCell(cell_pos, cell_type, cell_data, cell_style);
-		state = State::ROW;
-	}
-	else if(state == State::V && MatchTag("v", name)) {
-		state = State::CELL;
-		EnableTextHandler(false);
-	}
-	else if(state == State::IS && MatchTag("is", name)) {
-		state = State::CELL;
-	}
-	else if(state == State::T && MatchTag("t", name)) {
-		state = State::IS;
-		EnableTextHandler(false);
-	}
-}
-
-//-------------------------------------------------------------------
-// Range Sniffer
-//-------------------------------------------------------------------
-// The range sniffer is used to determine the range of the sheet to scan
-class RangeSniffer final : public SheetParserBase {
-public:
-	XLSXCellRange GetRange() const;
-private:
-	void OnEndRow(idx_t row_idx) override;
-	void OnCell(const XLSXCellPos &pos, XLSXCellType type, vector<char> &data, idx_t style) override;
-private:
-	// The range of the first consecutive non-empty row
-	idx_t beg_col = 0;
-	idx_t end_col = 0;
-	idx_t beg_row = 0;
-
-	// The total range of the sheet
-	idx_t min_col = 0;
-	idx_t max_col = 0;
-
-	enum class State : uint8_t { EMPTY, FOUND, ENDED};
-	State state = State::EMPTY;
-};
-
-XLSXCellRange RangeSniffer::GetRange() const {
-	if(beg_row == 0) {
-		// We didnt find any rows... return the whole sheet
-		return XLSXCellRange(1, 1, XLSX_MAX_CELL_ROWS, XLSX_MAX_CELL_COLS);
-	}
-	// Otherwise, return the sniffed range
-	return XLSXCellRange(beg_row, beg_col, XLSX_MAX_CELL_ROWS, end_col + 1);
-}
-
-void RangeSniffer::OnCell(const XLSXCellPos &pos, XLSXCellType type, vector<char> &data, idx_t style) {
-	min_col = std::min(min_col, pos.col);
-	max_col = std::max(max_col, pos.col);
-
-	switch(state) {
-		case State::EMPTY:
-			if(!data.empty()) {
-				state = State::FOUND;
-				beg_col = pos.col;
-				end_col = pos.col;
-			}
-		break;
-		case State::FOUND:
-			if(data.empty()) {
-				state = State::ENDED;
-			} else {
-				end_col = pos.col;
-			}
-		break;
-		case State::ENDED:
-			// We're done
-			break;
-	}
-}
-
-void RangeSniffer::OnEndRow(const idx_t row_idx) {
-	if(state == State::FOUND || state == State::ENDED) {
-		// We found a row with data, between beg_col and end_col
-		// We can now use this as the range for the sheet
-		beg_row = row_idx;
-		Stop(false);
-	} else {
-		// Reset, continue on to the next row
-		state = State::EMPTY;
-		beg_col = 0;
-		end_col = 0;
-	}
-}
-
-//-------------------------------------------------------------------
-// Header Sniffer
-//-------------------------------------------------------------------
-// The header sniffer is used to determine the header and the types
-// of the columns in the sheet (within the range)
-
-class HeaderSniffer final : public SheetParserBase {
-public:
-	HeaderSniffer(const XLSXCellRange &range_p, const XLSXHeaderMode header_mode_p, const bool absolute_range_p,
-		XLSXCellType default_cell_type_p)
-		: range(range_p), header_mode(header_mode_p), absolute_range(absolute_range_p), default_cell_type(default_cell_type_p) {
-	}
-
-	const XLSXCellRange &GetRange() const { return range; }
-	vector<XLSXCell> &GetColumnCells() { return column_cells; }
-	vector<XLSXCell> &GetHeaderCells() { return header_cells; }
-
-private:
-	void OnBeginRow(idx_t row_idx) override;
-	void OnEndRow(idx_t row_idx) override;
-	void OnCell(const XLSXCellPos &pos, XLSXCellType type, vector<char> &data, idx_t style) override;
-private:
-	vector<XLSXCell> header_cells;
-	vector<XLSXCell> column_cells;
-
-	XLSXCellRange range;
-	XLSXHeaderMode header_mode;
-
-	idx_t last_col = 0;
-
-	bool first_row = true;
-	bool absolute_range;
-	XLSXCellType default_cell_type;
-};
-
-void HeaderSniffer::OnBeginRow(const idx_t row_idx) {
-	if(!range.ContainsRow(row_idx)) {
-		return;
-	}
-	column_cells.clear();
-	last_col = range.beg.col - 1;
-}
-
-void HeaderSniffer::OnCell(const XLSXCellPos &pos, XLSXCellType type, vector<char> &data, idx_t style) {
-	if(!range.ContainsCol(pos.col)) {
-		return;
-	}
-
-	// Now, add the cell to the data cells, but make sure to pad with empty varchars if needed.
-	if(last_col + 1 < pos.col) {
-		// Pad with empty cells
-		for(idx_t i = last_col + 1; i < pos.col; i++) {
-			column_cells.emplace_back(default_cell_type, XLSXCellPos(pos.row, i), "", 0);
-		}
-	}
-
-	// Add the cell
-	column_cells.emplace_back(type, pos, string(data.data(), data.size()), style);
-	last_col = pos.col;
-}
-
-
-void HeaderSniffer::OnEndRow(const idx_t row_idx) {
-	if(!range.ContainsRow(row_idx)) {
-		column_cells.clear();
-		last_col = range.beg.col - 1;
-		return;
-	}
-
-	// If there are columns missing at the end, pad with empty string cells
-	if(last_col + 1 < range.end.col) {
-		for(idx_t i = last_col + 1; i < range.end.col; i++) {
-			column_cells.emplace_back(default_cell_type, XLSXCellPos(row_idx, i), "", 0);
-		}
-	}
-
-	// Now we have all the cells in the row, we can inspect them
-	if(!first_row) {
-		// This is the data row. We can stop here
-		Stop(false);
-		return;
-	}
-
-	// Now its time to determine the header row
-	auto has_header = false;
-	switch(header_mode) {
-		case XLSXHeaderMode::NEVER:
-			// We're not looking for a header, so we're done
-			has_header = false;
-			break;
-		case XLSXHeaderMode::FORCE:
-			has_header = true;
-			break;
-		case XLSXHeaderMode::MAYBE: {
-			// We're looking for a header, but we're not sure if we found it yet
-			// We need to inspect the cells to determine if this is a header row or not
-			// We need all the rows to be non empty and strings
-			auto all_strings = true;
-			for(const auto &cell : column_cells) {
-				auto &type = cell.type;
-				const auto is_str = type == XLSXCellType::SHARED_STRING || type == XLSXCellType::INLINE_STRING;
-				const auto is_empty = cell.data.empty();
-				if(!is_str || is_empty) {
-					all_strings = false;
-					break;
-				}
-			}
-			has_header = all_strings;
-		}
-	}
-
-	if(!has_header) {
-		// Generate a dummy header
-		header_cells = column_cells;
-		for(auto &cell : header_cells) {
-			cell.type = XLSXCellType::INLINE_STRING;
-			cell.style = 0;
-			if(!absolute_range) {
-				cell.data = cell.cell.ToString();
-			} else {
-				cell.data = cell.cell.GetColumnName();
-			}
-		}
-		Stop(false);
-		return;
-	}
-
-	// Save the header cells
-	header_cells = column_cells;
-	column_cells.clear();
-	last_col = range.beg.col - 1;
-
-	// Try to parse another row to see if we can find the data row
-	first_row = false;
-
-	// Move the range down one row
-	range.beg.row = row_idx + 1;
-}
-
-
-//-------------------------------------------------------------------
-// Sheet Parser
-//-------------------------------------------------------------------
-
-class SheetParser final : public SheetParserBase {
-public:
-	explicit SheetParser(ClientContext &context, const XLSXCellRange &range_p, const StringTable &table, bool stop_at_empty_p)
-		: string_table(table), range(range_p), stop_at_empty(stop_at_empty_p) {
-
-		// Initialize the chunk
-		const vector<LogicalType> types(range.Width(), LogicalType::VARCHAR);
-		auto &buffer_alloc = BufferAllocator::Get(context);
-		chunk.Initialize(buffer_alloc, types);
-
-		// Set the beginning column
-		// Allocate the sheet row number mapping
-		sheet_row_number = make_unsafe_uniq_array<idx_t>(STANDARD_VECTOR_SIZE);
-
-		last_row = range.beg.row - 1;
-		curr_row = range.beg.row;
-		last_col = range.beg.col - 1;
-	}
-
-	DataChunk& GetChunk() { return chunk; }
-	string GetCellName(idx_t chunk_row, idx_t chunk_col) const;
-
-	// Returns true if the chunk is full
-	bool FoundSkippedRow() const;
-	void SkipRows();
-	// Fill empty rows to the end of the range
-	void FillRows();
-
-protected:
-	void OnBeginRow(idx_t row_idx) override;
-	void OnEndRow(idx_t row_idx) override;
-	void OnCell(const XLSXCellPos &pos, XLSXCellType type, vector<char> &data, idx_t style) override;
-private:
-	// Shared String Table
-	const StringTable &string_table;
-	// Range to read
-	XLSXCellRange range;
-	// Mapping from chunk row to sheet row
-	unsafe_unique_array<idx_t> sheet_row_number;
-	// Current chunk
-	DataChunk chunk;
-	// Current row in the chunk
-	idx_t out_index = 0;
-
-	// The last column we wrote to
-	idx_t last_col;
-	// The last row we wrote to
-	idx_t last_row;
-	idx_t curr_row;
-
-	bool stop_at_empty = false;
-	bool is_row_empty = false;
-};
-
-string SheetParser::GetCellName(idx_t chunk_row, idx_t chunk_col) const {
-	// Get the cell name and row given a chunk row and column
-	const auto sheet_row = sheet_row_number[chunk_row];
-	const auto sheet_col = chunk_col + range.beg.col;
-
-	const XLSXCellPos pos = {static_cast<idx_t>(sheet_row), sheet_col};
-	return pos.ToString();
-}
-
-bool SheetParser::FoundSkippedRow() const {
-	return last_row + 1 < curr_row;
-}
-
-void SheetParser::SkipRows() {
-	// Pad empty rows
-	while (last_row + 1 < curr_row) {
-		last_row++;
-
-		for(auto &col : chunk.data) {
-			FlatVector::SetNull(col, out_index, true);
-		}
-		sheet_row_number[out_index] = last_row;
-		out_index++;
-		chunk.SetCardinality(out_index);
-
-		if(out_index == STANDARD_VECTOR_SIZE) {
-			// We have filled up the chunk, yield!
-			out_index = 0;
-			return;
-		}
-	}
-}
-
-void SheetParser::FillRows() {
-	const auto total_remaining = range.end.row - 1 - last_row;
-	const auto local_remaining = STANDARD_VECTOR_SIZE - out_index;
-
-	const auto remaining = MinValue(total_remaining, local_remaining);
-	for(idx_t i = 0; i < remaining; i++) {
-		for(auto &col : chunk.data) {
-			FlatVector::SetNull(col, out_index, true);
-		}
-		sheet_row_number[out_index] = last_row;
-		out_index++;
-		last_row++;
-	}
-	chunk.SetCardinality(chunk.size() + remaining);
-	out_index = 0;
-}
-
-
-void SheetParser::OnBeginRow(idx_t row_idx) {
-	if(!range.ContainsRow(row_idx)) {
-		// not in range, skip
-		return;
-	}
-
-	last_col = range.beg.col - 1;
-	is_row_empty = true;
-
-	curr_row = row_idx;
-
-	// Check if we need to pad empty rows
-	if(last_row + 1 < curr_row) {
-		Stop(true);
-	}
-}
-
-void SheetParser::OnCell(const XLSXCellPos &pos, XLSXCellType type, vector<char> &data, idx_t style) {
-	if(!range.ContainsPos(pos)) {
-		// not in range, skip
-		return;
-	}
-
-	// If we jumped over some columns, pad with nulls
-	if(last_col + 1 < pos.col) {
-		for(idx_t i = last_col + 1; i < pos.col; i++) {
-			auto &vec = chunk.data[i - range.beg.col];
-			FlatVector::SetNull(vec, out_index, true);
-		}
-	}
-
-	// Get the column data
-	auto &vec = chunk.data[pos.col - range.beg.col];
-
-	// Push the cell data to our chunk
-	const auto ptr = FlatVector::GetData<string_t>(vec);
-
-	if(type == XLSXCellType::SHARED_STRING) {
-		// Push a null to the buffer so that the string is null-terminated
-		data.push_back('\0');
-		// Now we can use strtol to get the shared string index
-		const auto ssi = std::strtol(data.data(), nullptr, 10);
-		// Look up the string in the string table
-		ptr[out_index] = string_table.Get(ssi);
-	}
-	else if(data.empty() && type != XLSXCellType::INLINE_STRING) {
-		// If the cell is empty (and not a string), we wont be able to convert it
-		// so just null it immediately
-		FlatVector::SetNull(vec, out_index, true);
-	} else {
-		// Otherwise just pass along the call data, we will cast it later.
-		ptr[out_index] = StringVector::AddString(vec, data.data(), data.size());
-	}
-
-	if(!data.empty()) {
-		is_row_empty = false;
-	}
-
-	last_col = pos.col;
-}
-
-
-void SheetParser::OnEndRow(idx_t row_idx) {
-	if(!range.ContainsRow(row_idx)) {
-		// not in range, skip
-		return;
-	}
-
-	last_row = row_idx;
-
-	if(stop_at_empty && is_row_empty) {
-		Stop(false);
-		return;
-	}
-
-	// If we didnt write out all the columns, pad with nulls
-	if(last_col + 1 < range.end.col) {
-		for(idx_t i = last_col + 1; i < range.end.col; i++) {
-			auto &vec = chunk.data[i - range.beg.col];
-			FlatVector::SetNull(vec, out_index, true);
-		}
-	}
-
-	// Map the chunk row to the sheet row
-	sheet_row_number[out_index] = UnsafeNumericCast<int32_t>(row_idx);
-
-	out_index++;
-	chunk.SetCardinality(out_index);
-	if(out_index == STANDARD_VECTOR_SIZE) {
-		// We have filled up the chunk, yield!
-		out_index = 0;
-		Stop(true);
-	}
-}
 
 //-------------------------------------------------------------------
 // Meta
@@ -1235,7 +126,7 @@ static void ResolveColumnNames(vector<XLSXCell> &header_cells, ZipFileReader &ar
 	if(!archive.TryOpenEntry("xl/sharedStrings.xml")) {
 		throw BinderException("No shared strings found in xlsx file");
 	}
-	StringTableSearcher searcher(shared_string_ids);
+	SharedStringSearcher searcher(shared_string_ids);
 	searcher.ParseAll(archive);
 	archive.CloseEntry();
 
@@ -1247,9 +138,6 @@ static void ResolveColumnNames(vector<XLSXCell> &header_cells, ZipFileReader &ar
 	}
 }
 
-//-------------------------------------------------------------------
-// Bind
-//-------------------------------------------------------------------
 
 void ReadXLSX::ParseOptions(XLSXReadOptions &options, const named_parameter_map_t &input) {
 
@@ -1294,15 +182,14 @@ void ReadXLSX::ParseOptions(XLSXReadOptions &options, const named_parameter_map_
 		range.end.row++;
 
 		options.range = range;
-		options.stop_at_empty = false;
 		options.has_explicit_range = true;
+
+		// Default to not stop at empty if a range is specified
+		options.stop_at_empty = false;
 	}
 
 	const auto stop_at_empty_op = input.find("stop_at_empty");
 	if(stop_at_empty_op != input.end()) {
-		if(options.has_explicit_range) {
-			throw BinderException("Cannot specify both 'range' and 'stop_at_empty' options");
-		}
 		options.stop_at_empty = BooleanValue::Get(stop_at_empty_op->second);
 	}
 
@@ -1396,6 +283,10 @@ void ReadXLSX::ResolveSheet(const unique_ptr<XLSXReadData> &result, ZipFileReade
 	SniffHeader(result, archive);
 }
 
+//-------------------------------------------------------------------
+// Bind
+//-------------------------------------------------------------------
+
 static unique_ptr<FunctionData> Bind(ClientContext &context, TableFunctionBindInput &input,
 														  vector<LogicalType> &return_types, vector<string> &names) {
 	auto result = make_uniq<XLSXReadData>();
@@ -1454,7 +345,7 @@ static unique_ptr<GlobalTableFunctionState> InitGlobal(ClientContext &context, T
 
 	// Check if there is a string table. If there is, extract it
 	if(state->archive.TryOpenEntry("xl/sharedStrings.xml")) {
-		StringTableParser::ParseStringTable(state->archive, state->strings);
+		SharedStringParser::ParseStringTable(state->archive, state->strings);
 		state->archive.CloseEntry();
 	}
 
@@ -1645,12 +536,7 @@ static double Progress(ClientContext &context, const FunctionData *bind_data_p,
 	const auto pos = static_cast<double>(state.stream_pos.load());
 	const auto len = static_cast<double>(state.stream_len);
 
-	// Safety check
-	if(pos == 0 || len == 0) {
-		return 0;
-	}
-
-	return pos / len;
+	return (pos == 0 || len == 0) ? 0 : (pos / len) * 100.0;
 }
 
 static unique_ptr<TableRef> XLSXReplacementScan(ClientContext &context, ReplacementScanInput &input,
